blog/users/utils.py
'''
Author: yuan
Date: 2021-05-24 13:42:22
LastEditTime: 2021-05-24 16:36:27
FilePath: /python-flask/blog/users/utils.py
'''
import os
import secrets
from PIL import Image
from flask import url_for, current_app
from flask_mail import Message
from blog import mail


def save_picture(form_picture):
    random_hex = secrets.token_hex(8)
    f_name, f_ext = os.path.splitext(form_picture.filename)
    picture_fn = random_hex + f_ext
    picture_path = os.path.join(current_app.root_path, "static/profile_pic", picture_fn)

    output_size = (128, 128)
    i = Image.open(form_picture)
    i.thumbnail(output_size)  # image compression
    i.save(picture_path)

    return picture_fn


def send_reset_email(user):
    token = user.get_reset_token()
    msg = Message(
        "Password Reset Request", sender="a920604a@gmail.com", recipients=[user.email]
    )
    msg.body = f'''To reset your password, visit the following link:
    {url_for('users.reset_token', token=token, _external=True)}
    If you did not make thie request then simply ignore this email and no change'''
    mail.send(msg)