requirements.txt
Flask
Flask-WTF
Flask-SQLAlchemy
Flask-Bcrypt
Flask-Login
Pillow
itsdangerous
Flask-Mail
WTForms
SQLAlchemy
bcrypt
email_validator
Flask-markdown 
Flask-PageDown
click
Jinja2
MarkupSafe
Werkzeug
markdown
